Audit Log: Prototype - read entries from file
The /tmp/auditLogSample.json file is read to obtain the Audit Log
entries for:
  $bmc/redfish/v1/Systems/system/LogServices/AuditLog/Entries

The file contains a single JSon object per line with the details of the
audit log entry. This is the format the D-Bus audit log parser method
will produce and send the unixfd to bmcweb to process.

Each object is expected to contain each of these elements:
  - "ID"
  - "EventTimestamp"
  - "Message"
  - "MessageArgs"

If the object is malformed a debug message will be issued but the
collection will just skip the bad object and return any of the objects
that are good. This is for the prototype only and will change before
final code.

This code is quite similar to the eventual code, but it won't be
looking in the fixed location for the file to parse.

This handles the top and skip query parameters, e.g.:
https://${BMC_IP}/redfish/v1/Systems/system/LogServices/AuditLog/Entries?\$skip=19
 https://${BMC_IP}/redfish/v1/Systems/system/LogServices/AuditLog/Entries?\$top=2\&\$skip=10

Also handles retrieving a specific entry, e.g.:
https://${BMC_IP}/redfish/v1/Systems/system/LogServices/AuditLog/Entries/1692305135.139:9

Tested:
  - Tried all of the new routes.
  - Sampled retrieving specific entry and non-existent entry.
  - Tried various combinations with top and skip query parameters.
  - Redfish Validator: No new errors, passes AuditLog related tests.

Signed-off-by: Janet Adkins <janeta@us.ibm.com>